change flink-operator only in jdk11
