+++ dlink-admin/src/main/java/com/dlink/service/impl/TaskServiceImpl.java
@@ -41,7 +41,9 @@ import com.dlink.dto.TaskVersionConfigureDTO;
 import com.dlink.exception.BusException;
 import com.dlink.function.compiler.CustomStringJavaCompiler;
 import com.dlink.function.util.UDFUtil;
+import com.dlink.gateway.Gateway;
 import com.dlink.gateway.GatewayType;
+import com.dlink.gateway.config.GatewayConfig;
 import com.dlink.gateway.config.SavePointStrategy;
 import com.dlink.gateway.config.SavePointType;
 import com.dlink.gateway.model.JobInfo;
@@ -198,7 +200,7 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
 
     private String buildParas(Integer id) {
         return "--id " + id + " --driver " + driver + " --url " + url + " --username " + username + " --password "
-            + password;
+                + password;
     }
 
     @Override
@@ -207,7 +209,7 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
         Asserts.checkNull(task, Tips.TASK_NOT_EXIST);
         if (Dialect.notFlinkSql(task.getDialect())) {
             return executeCommonSql(SqlDTO.build(task.getStatement(),
-                task.getDatabaseId(), null));
+                    task.getDatabaseId(), null));
         }
         JobConfig config = buildJobConfig(task);
         // init UDF
@@ -228,7 +230,7 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
         task.setStep(JobLifeCycle.ONLINE.getValue());
         if (Dialect.notFlinkSql(task.getDialect())) {
             return executeCommonSql(SqlDTO.build(task.getStatement(),
-                task.getDatabaseId(), null));
+                    task.getDatabaseId(), null));
         }
         JobConfig config = buildJobConfig(task);
         JobManager jobManager = JobManager.build(config);
@@ -248,7 +250,7 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
         }
         if (Dialect.notFlinkSql(task.getDialect())) {
             return executeCommonSql(SqlDTO.build(task.getStatement(),
-                task.getDatabaseId(), null));
+                    task.getDatabaseId(), null));
         }
         if (StringUtils.isBlank(savePointPath)) {
             task.setSavePointStrategy(SavePointStrategy.LATEST.getValue());
@@ -373,13 +375,13 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
                 UDFTemplate template = udfTemplateService.getById(config.get("templateId"));
                 if (template != null) {
                     String code = UDFUtil.templateParse(task.getDialect(), template.getTemplateCode(),
-                        config.get("className"));
+                            config.get("className"));
                     task.setStatement(code);
                 }
             }
             // to compiler udf
             if (Asserts.isNotNullString(task.getDialect()) && Dialect.JAVA.equalsVal(task.getDialect())
-                && Asserts.isNotNullString(task.getStatement())) {
+                    && Asserts.isNotNullString(task.getStatement())) {
                 CustomStringJavaCompiler compiler = new CustomStringJavaCompiler(task.getStatement());
                 task.setSavePointPath(compiler.getFullClassName());
             } else if (Dialect.PYTHON.equalsVal(task.getDialect())) {
@@ -394,8 +396,8 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
             Task taskInfo = getById(task.getId());
             Assert.check(taskInfo);
             if (JobLifeCycle.RELEASE.equalsValue(taskInfo.getStep())
-                || JobLifeCycle.ONLINE.equalsValue(taskInfo.getStep())
-                || JobLifeCycle.CANCEL.equalsValue(taskInfo.getStep())) {
+                    || JobLifeCycle.ONLINE.equalsValue(taskInfo.getStep())
+                    || JobLifeCycle.CANCEL.equalsValue(taskInfo.getStep())) {
                 throw new BusException("该作业已" + JobLifeCycle.get(taskInfo.getStep()).getLabel() + "，禁止修改！");
             }
             task.setStep(JobLifeCycle.DEVELOP.getValue());
@@ -488,6 +490,19 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
     }
 
     @Override
+    public JobStatus checkJobStatus(JobInfoDetail jobInfoDetail) {
+        JobConfig jobConfig = new JobConfig();
+        Map<String, Object> gatewayConfigMap =
+                clusterConfigurationService.getGatewayConfig(jobInfoDetail.getClusterConfiguration().getId());
+        jobConfig.buildGatewayConfig(gatewayConfigMap);
+        GatewayConfig gatewayConfig = jobConfig.getGatewayConfig();
+        gatewayConfig.setType(GatewayType.get(jobInfoDetail.getCluster().getType()));
+        gatewayConfig.getClusterConfig().setAppId(jobInfoDetail.getCluster().getName());
+        Gateway gateway = Gateway.build(gatewayConfig);
+        return gateway.getJobStatusById(jobInfoDetail.getCluster().getName());
+    }
+
+    @Override
     public String exportSql(Integer id) {
         Task task = getTaskInfoById(id);
         Asserts.checkNull(task, Tips.TASK_NOT_EXIST);
@@ -506,8 +521,8 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
     @Override
     public Task getUDFByClassName(String className) {
         Task task = getOne(
-            new QueryWrapper<Task>().in("dialect", Dialect.JAVA, Dialect.SCALA, Dialect.PYTHON).eq("enabled", 1)
-                .eq("save_point_path", className));
+                new QueryWrapper<Task>().in("dialect", Dialect.JAVA, Dialect.SCALA, Dialect.PYTHON).eq("enabled", 1)
+                        .eq("save_point_path", className));
         Assert.check(task);
         task.setStatement(statementService.getById(task.getId()).getStatement());
         return task;
@@ -516,8 +531,8 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
     @Override
     public List<Task> getAllUDF() {
         List<Task> tasks =
-            list(new QueryWrapper<Task>().in("dialect", Dialect.JAVA, Dialect.SCALA, Dialect.PYTHON)
-                .eq("enabled", 1).isNotNull("save_point_path"));
+                list(new QueryWrapper<Task>().in("dialect", Dialect.JAVA, Dialect.SCALA, Dialect.PYTHON)
+                        .eq("enabled", 1).isNotNull("save_point_path"));
         return tasks.stream().peek(task -> {
             Assert.check(task);
             task.setStatement(statementService.getById(task.getId()).getStatement());
@@ -554,7 +569,7 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
         List<TaskVersion> taskVersions = taskVersionService.getTaskVersionByTaskId(task.getId());
         List<Integer> versionIds = taskVersions.stream().map(TaskVersion::getVersionId).collect(Collectors.toList());
         Map<Integer, TaskVersion> versionMap =
-            taskVersions.stream().collect(Collectors.toMap(TaskVersion::getVersionId, t -> t));
+                taskVersions.stream().collect(Collectors.toMap(TaskVersion::getVersionId, t -> t));
 
         TaskVersion taskVersion = new TaskVersion();
         BeanUtil.copyProperties(task, taskVersion);
@@ -593,14 +608,14 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
         Task taskInfo = getTaskInfoById(dto.getId());
 
         if (JobLifeCycle.RELEASE.equalsValue(taskInfo.getStep())
-            || JobLifeCycle.ONLINE.equalsValue(taskInfo.getStep())
-            || JobLifeCycle.CANCEL.equalsValue(taskInfo.getStep())) {
+                || JobLifeCycle.ONLINE.equalsValue(taskInfo.getStep())
+                || JobLifeCycle.CANCEL.equalsValue(taskInfo.getStep())) {
             // throw new BusException("该作业已" + JobLifeCycle.get(taskInfo.getStep()).getLabel() + "，禁止回滚！");
             return Result.failed("该作业已" + JobLifeCycle.get(taskInfo.getStep()).getLabel() + "，禁止回滚！");
         }
 
         LambdaQueryWrapper<TaskVersion> queryWrapper = new LambdaQueryWrapper<TaskVersion>()
-            .eq(TaskVersion::getTaskId, dto.getId()).eq(TaskVersion::getVersionId, dto.getVersionId());
+                .eq(TaskVersion::getTaskId, dto.getId()).eq(TaskVersion::getVersionId, dto.getVersionId());
 
         TaskVersion taskVersion = taskVersionService.getOne(queryWrapper);
 
@@ -747,7 +762,7 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
         }
         if (Asserts.isNotNull(cluster.getClusterConfigurationId())) {
             Map<String, Object> gatewayConfig =
-                clusterConfigurationService.getGatewayConfig(cluster.getClusterConfigurationId());
+                    clusterConfigurationService.getGatewayConfig(cluster.getClusterConfigurationId());
             jobConfig.buildGatewayConfig(gatewayConfig);
             jobConfig.getGatewayConfig().getClusterConfig().setAppId(cluster.getName());
             useGateway = true;
@@ -783,7 +798,7 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
 
     private JobConfig buildJobConfig(Task task) {
         boolean isJarTask = Dialect.FLINKJAR.equalsVal(task.getDialect())
-            || Dialect.KUBERNETES_APPLICATION.equalsVal(task.getDialect());
+                || Dialect.KUBERNETES_APPLICATION.equalsVal(task.getDialect());
         if (!isJarTask && Asserts.isNotNull(task.getFragment()) ? task.getFragment() : false) {
             String flinkWithSql = dataBaseService.getEnabledFlinkWithSql();
             if (Asserts.isNotNullString(flinkWithSql)) {
@@ -804,7 +819,7 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
         // support custom K8s app submit, rather than clusterConfiguration
         else if (Dialect.KUBERNETES_APPLICATION.equalsVal(task.getDialect())
                 && (GatewayType.KUBERNETES_APPLICATION.equalsValue(config.getType())
-                || GatewayType.KUBERNETES_APPLICATION_OPERATOR.equalsValue(config.getType()))) {
+                        || GatewayType.KUBERNETES_APPLICATION_OPERATOR.equalsValue(config.getType()))) {
             Map<String, Object> taskConfig = JSONUtil.toMap(task.getStatement(), String.class, Object.class);
             Map<String, Object> clusterConfiguration =
                     clusterConfigurationService.getGatewayConfig(task.getClusterConfigurationId());
@@ -813,7 +828,7 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
             config.buildGatewayConfig(clusterConfiguration);
         } else {
             Map<String, Object> gatewayConfig =
-                clusterConfigurationService.getGatewayConfig(task.getClusterConfigurationId());
+                    clusterConfigurationService.getGatewayConfig(task.getClusterConfigurationId());
             // submit application type with clusterConfiguration
             if (GatewayType.YARN_APPLICATION.equalsValue(config.getType())
                     || GatewayType.KUBERNETES_APPLICATION.equalsValue(config.getType())
@@ -822,7 +837,7 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
                     SystemConfiguration systemConfiguration = SystemConfiguration.getInstances();
                     gatewayConfig.put("userJarPath", systemConfiguration.getSqlSubmitJarPath());
                     gatewayConfig.put("userJarParas",
-                        systemConfiguration.getSqlSubmitJarParas() + buildParas(config.getTaskId()));
+                            systemConfiguration.getSqlSubmitJarParas() + buildParas(config.getTaskId()));
                     gatewayConfig.put("userJarMainAppClass", systemConfiguration.getSqlSubmitJarMainAppClass());
                 } else {
                     Jar jar = jarService.getById(task.getJarId());
@@ -858,6 +873,7 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
                 config.setSavePointPath(null);
         }
         config.setVariables(fragmentVariableService.listEnabledVariables());
+        config.getGatewayConfig().getAppConfig().setParallelism(task.getParallelism());
         return config;
     }
 
@@ -937,7 +953,7 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
 
     private boolean inRefreshPlan(JobInstance jobInstance) {
         if ((!JobStatus.isDone(jobInstance.getStatus())) || (Asserts.isNotNull(jobInstance.getFinishTime())
-            && Duration.between(jobInstance.getFinishTime(), LocalDateTime.now()).toMinutes() < 1)) {
+                && Duration.between(jobInstance.getFinishTime(), LocalDateTime.now()).toMinutes() < 1)) {
             return true;
         } else {
             return false;
@@ -984,9 +1000,9 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
         // clusterConfigurationName
         if (Asserts.isNotNull(task.getClusterConfigurationId())) {
             ClusterConfiguration clusterConfiguration =
-                clusterConfigurationService.getById(task.getClusterConfigurationId());
+                    clusterConfigurationService.getById(task.getClusterConfigurationId());
             ((ObjectNode) jsonNode).put("clusterConfigurationName",
-                Asserts.isNotNull(clusterConfiguration) ? clusterConfiguration.getName() : null);
+                    Asserts.isNotNull(clusterConfiguration) ? clusterConfiguration.getName() : null);
         }
         // databaseName
         if (Asserts.isNotNull(task.getDatabaseId())) {
@@ -1062,15 +1078,15 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
             }
             if (Asserts.isNotNull(task.getClusterConfigurationName())) {
                 ClusterConfiguration clusterConfiguration = clusterConfigurationService
-                    .getOne(new QueryWrapper<ClusterConfiguration>().eq("name",
-                        task.getClusterConfigurationName()));
+                        .getOne(new QueryWrapper<ClusterConfiguration>().eq("name",
+                                task.getClusterConfigurationName()));
                 if (Asserts.isNotNull(clusterConfiguration)) {
                     task.setClusterConfigurationId(clusterConfiguration.getId());
                 }
             }
             if (Asserts.isNotNull(task.getDatabaseName())) {
                 DataBase dataBase =
-                    dataBaseService.getOne(new QueryWrapper<DataBase>().eq("name", task.getDatabaseName()));
+                        dataBaseService.getOne(new QueryWrapper<DataBase>().eq("name", task.getDatabaseName()));
                 if (Asserts.isNotNull(dataBase)) {
                     task.setDatabaseId(dataBase.getId());
                 }
@@ -1087,7 +1103,7 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
              */
             if (Asserts.isNotNull(task.getAlertGroupName())) {
                 AlertGroup alertGroup =
-                    alertGroupService.getOne(new QueryWrapper<AlertGroup>().eq("name", task.getAlertGroupName()));
+                        alertGroupService.getOne(new QueryWrapper<AlertGroup>().eq("name", task.getAlertGroupName()));
                 if (Asserts.isNotNull(alertGroup)) {
                     task.setAlertGroupId(alertGroup.getId());
                 }
@@ -1123,7 +1139,7 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
             return Result.failed("一共" + jsonNodes.size() + "个作业,全部导入失败");
         } else if (errorNumber > 0) {
             return Result.failed("一共" + jsonNodes.size() + "个作业,其中成功导入" + (jsonNode.size() - errorNumber) + "个,失败"
-                + errorNumber + "个");
+                    + errorNumber + "个");
         }
         return Result.succeed("成功导入" + jsonNodes.size() + "个作业");
     }
@@ -1174,7 +1190,7 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
         long minutes = ChronoUnit.MINUTES.between(startTime, endTime);
         long seconds = ChronoUnit.SECONDS.between(startTime, endTime);
         String duration = days + "天 " + (hours - (days * 24)) + "小时 " + (minutes - (hours * 60)) + "分 "
-            + (seconds - (minutes * 60)) + "秒";
+                + (seconds - (minutes * 60)) + "秒";
         return duration;
     }
 
@@ -1254,7 +1270,7 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
 
     private void sendAlert(AlertInstance alertInstance, JobInstance jobInstance, Task task, AlertMsg alertMsg) {
         AlertConfig alertConfig = AlertConfig.build(alertInstance.getName(), alertInstance.getType(),
-            JSONUtil.toMap(alertInstance.getParams()));
+                JSONUtil.toMap(alertInstance.getParams()));
         Alert alert = Alert.build(alertConfig);
         String title = "任务【" + task.getAlias() + "】：" + jobInstance.getStatus();
         String content = alertMsg.toString();
@@ -1273,10 +1289,10 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
     @Override
     public Result queryAllCatalogue() {
         final LambdaQueryWrapper<Catalogue> queryWrapper = new LambdaQueryWrapper<Catalogue>()
-            .select(Catalogue::getId, Catalogue::getName, Catalogue::getParentId)
-            .eq(Catalogue::getIsLeaf, 0)
-            .eq(Catalogue::getEnabled, 1)
-            .isNull(Catalogue::getTaskId);
+                .select(Catalogue::getId, Catalogue::getName, Catalogue::getParentId)
+                .eq(Catalogue::getIsLeaf, 0)
+                .eq(Catalogue::getEnabled, 1)
+                .isNull(Catalogue::getTaskId);
         final List<Catalogue> catalogueList = catalogueService.list(queryWrapper);
         return Result.succeed(TreeUtil.build(dealWithCatalogue(catalogueList), -1).get(0));
     }
@@ -1302,7 +1318,7 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
     public Result<List<Task>> queryOnLineTaskByDoneStatus(List<JobLifeCycle> jobLifeCycle, List<JobStatus> jobStatuses,
                                                           boolean includeNull, Integer catalogueId) {
         final Tree<Integer> node = ((Tree<Integer>) queryAllCatalogue().getDatas())
-            .getNode(Objects.isNull(catalogueId) ? 0 : catalogueId);
+                .getNode(Objects.isNull(catalogueId) ? 0 : catalogueId);
         final List<Integer> parentIds = new ArrayList<>(0);
         parentIds.add(node.getId());
         childrenNodeParse(node, parentIds);
@@ -1313,8 +1329,8 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
     private List<Task> getTasks(List<JobLifeCycle> jobLifeCycle, List<JobStatus> jobStatuses, boolean includeNull,
                                 List<Integer> parentIds) {
         return this.baseMapper.queryOnLineTaskByDoneStatus(parentIds,
-            jobLifeCycle.stream().filter(Objects::nonNull).map(JobLifeCycle::getValue).collect(Collectors.toList()),
-            includeNull, jobStatuses.stream().map(JobStatus::name).collect(Collectors.toList()));
+                jobLifeCycle.stream().filter(Objects::nonNull).map(JobLifeCycle::getValue).collect(Collectors.toList()),
+                includeNull, jobStatuses.stream().map(JobStatus::name).collect(Collectors.toList()));
     }
 
     private void childrenNodeParse(Tree<Integer> node, List<Integer> parentIds) {
@@ -1334,7 +1350,7 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
     @Override
     public void selectSavepointOnLineTask(TaskOperatingResult taskOperatingResult) {
         final JobInstance jobInstanceByTaskId =
-            jobInstanceService.getJobInstanceByTaskId(taskOperatingResult.getTask().getId());
+                jobInstanceService.getJobInstanceByTaskId(taskOperatingResult.getTask().getId());
         if (jobInstanceByTaskId == null) {
             startGoingLiveTask(taskOperatingResult, null);
             return;
@@ -1353,8 +1369,8 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
     private void findTheConditionSavePointToOnline(TaskOperatingResult taskOperatingResult,
                                                    JobInstance jobInstanceByTaskId) {
         final LambdaQueryWrapper<JobHistory> queryWrapper = new LambdaQueryWrapper<JobHistory>()
-            .select(JobHistory::getId, JobHistory::getCheckpointsJson)
-            .eq(JobHistory::getId, jobInstanceByTaskId.getId());
+                .select(JobHistory::getId, JobHistory::getCheckpointsJson)
+                .eq(JobHistory::getId, jobInstanceByTaskId.getId());
         final JobHistory jobHistory = jobHistoryService.getOne(queryWrapper);
         if (jobHistory != null && StringUtils.isNotBlank(jobHistory.getCheckpointsJson())) {
             final ObjectNode jsonNodes = JSONUtil.parseObject(jobHistory.getCheckpointsJson());
