+++ dlink-gateway/dlink-gateway-base/pom.xml
@@ -1,78 +1,70 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Licensed to the Apache Software Foundation (ASF) under one or more
-  ~ contributor license agreements.  See the NOTICE file distributed with
-  ~ this work for additional information regarding copyright ownership.
-  ~ The ASF licenses this file to You under the Apache License, Version 2.0
-  ~ (the "License"); you may not use this file except in compliance with
-  ~ the License.  You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>com.dlink</groupId>
-        <artifactId>dlink</artifactId>
+        <artifactId>dlink-gateway</artifactId>
         <version>0.6.8-SNAPSHOT</version>
     </parent>
 
-    <artifactId>dlink-function</artifactId>
+    <artifactId>dlink-gateway-base</artifactId>
+
+    <properties>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <maven.compiler.source>${target.java.version}</maven.compiler.source>
+        <maven.compiler.target>${target.java.version}</maven.compiler.target>
+    </properties>
 
     <dependencies>
         <dependency>
+            <groupId>com.dlink</groupId>
+            <artifactId>dlink-common</artifactId>
+        </dependency>
+
+        <dependency>
             <groupId>cn.hutool</groupId>
             <artifactId>hutool-all</artifactId>
-            <scope>provided</scope>
         </dependency>
         <dependency>
-            <groupId>com.dlink</groupId>
-            <artifactId>dlink-gateway-base</artifactId>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
             <scope>provided</scope>
         </dependency>
+
+        <!--        delete this dependency will result in a JcaPEMKey error
+        for k8s commit tasks-->
         <dependency>
-            <groupId>com.dlink</groupId>
-            <artifactId>dlink-process</artifactId>
-            <scope>provided</scope>
+            <groupId>org.bouncycastle</groupId>
+            <artifactId>bcpkix-jdk15on</artifactId>
+            <version>1.69</version>
         </dependency>
         <dependency>
-            <groupId>org.codehaus.groovy</groupId>
-            <artifactId>groovy</artifactId>
-            <scope>provided</scope>
+            <groupId>org.bouncycastle</groupId>
+            <artifactId>bcprov-jdk15on</artifactId>
+            <version>1.69</version>
         </dependency>
         <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-context</artifactId>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.scala-lang</groupId>
-            <artifactId>scala-compiler</artifactId>
-            <version>${scala.version}</version>
-            <scope>provided</scope>
+            <groupId>org.bouncycastle</groupId>
+            <artifactId>bcprov-ext-jdk15on</artifactId>
+            <version>1.69</version>
         </dependency>
     </dependencies>
 
     <profiles>
+
         <profile>
             <id>flink-1.11</id>
             <dependencies>
                 <dependency>
                     <groupId>com.dlink</groupId>
                     <artifactId>dlink-client-1.11</artifactId>
-                    <scope>provided</scope>
+                    <scope>${scope.runtime}</scope>
                 </dependency>
                 <dependency>
                     <groupId>com.dlink</groupId>
                     <artifactId>dlink-flink-1.11</artifactId>
-                    <scope>provided</scope>
+                    <scope>${scope.runtime}</scope>
                 </dependency>
             </dependencies>
         </profile>
@@ -83,12 +75,12 @@
                 <dependency>
                     <groupId>com.dlink</groupId>
                     <artifactId>dlink-client-1.12</artifactId>
-                    <scope>provided</scope>
+                    <scope>${scope.runtime}</scope>
                 </dependency>
                 <dependency>
                     <groupId>com.dlink</groupId>
                     <artifactId>dlink-flink-1.12</artifactId>
-                    <scope>provided</scope>
+                    <scope>${scope.runtime}</scope>
                 </dependency>
             </dependencies>
         </profile>
@@ -99,12 +91,12 @@
                 <dependency>
                     <groupId>com.dlink</groupId>
                     <artifactId>dlink-client-1.13</artifactId>
-                    <scope>provided</scope>
+                    <scope>${scope.runtime}</scope>
                 </dependency>
                 <dependency>
                     <groupId>com.dlink</groupId>
                     <artifactId>dlink-flink-1.13</artifactId>
-                    <scope>provided</scope>
+                    <scope>${scope.runtime}</scope>
                 </dependency>
             </dependencies>
         </profile>
@@ -115,12 +107,12 @@
                 <dependency>
                     <groupId>com.dlink</groupId>
                     <artifactId>dlink-client-1.14</artifactId>
-                    <scope>provided</scope>
+                    <scope>${scope.runtime}</scope>
                 </dependency>
                 <dependency>
                     <groupId>com.dlink</groupId>
                     <artifactId>dlink-flink-1.14</artifactId>
-                    <scope>provided</scope>
+                    <scope>${scope.runtime}</scope>
                 </dependency>
             </dependencies>
         </profile>
@@ -131,12 +123,12 @@
                 <dependency>
                     <groupId>com.dlink</groupId>
                     <artifactId>dlink-client-1.15</artifactId>
-                    <scope>provided</scope>
+                    <scope>${scope.runtime}</scope>
                 </dependency>
                 <dependency>
                     <groupId>com.dlink</groupId>
                     <artifactId>dlink-flink-1.15</artifactId>
-                    <scope>provided</scope>
+                    <scope>${scope.runtime}</scope>
                 </dependency>
             </dependencies>
         </profile>
@@ -147,14 +139,15 @@
                 <dependency>
                     <groupId>com.dlink</groupId>
                     <artifactId>dlink-client-1.16</artifactId>
-                    <scope>provided</scope>
+                    <scope>${scope.runtime}</scope>
                 </dependency>
                 <dependency>
                     <groupId>com.dlink</groupId>
                     <artifactId>dlink-flink-1.16</artifactId>
-                    <scope>provided</scope>
+                    <scope>${scope.runtime}</scope>
                 </dependency>
             </dependencies>
         </profile>
     </profiles>
+
 </project>
