+++ dlink-gateway/dlink-gateway-base/src/main/java/com/dlink/gateway/GatewayType.java
