+++ dlink-gateway/dlink-gateway-base/src/main/java/com/dlink/gateway/exception/GatewayException.java
