+++ dlink-gateway/dlink-gateway-base/src/main/java/com/dlink/gateway/result/GatewayResult.java
