+++ dlink-gateway/dlink-gateway-kubernets/src/main/resources/META-INF/services/com.dlink.gateway.Gateway
@@ -1,0 +1,1 @@
+com.dlink.gateway.kubernetes.KubernetesApplicationGateway
