+++ dlink-gateway/dlink-gateway-yarn/src/main/resources/META-INF/services/com.dlink.gateway.Gateway
@@ -1,0 +1,2 @@
+com.dlink.gateway.yarn.YarnApplicationGateway
+com.dlink.gateway.yarn.YarnPerJobGateway
