--- dlink-gateway/src/main/java/com/dlink/gateway/config/SavePointType.java
@@ -1,51 +1,0 @@
-/*
- *
- *  Licensed to the Apache Software Foundation (ASF) under one or more
- *  contributor license agreements.  See the NOTICE file distributed with
- *  this work for additional information regarding copyright ownership.
- *  The ASF licenses this file to You under the Apache License, Version 2.0
- *  (the "License"); you may not use this file except in compliance with
- *  the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *
- */
-
-package com.dlink.gateway.config;
-
-import com.dlink.assertion.Asserts;
-
-/**
- * SavePointType
- *
- * @author wenmo
- * @since 2021/11/3 21:58
- */
-public enum SavePointType {
-    TRIGGER("trigger"), DISPOSE("dispose"), STOP("stop"), CANCEL("cancel");
-
-    private String value;
-
-    SavePointType(String value) {
-        this.value = value;
-    }
-
-    public String getValue() {
-        return value;
-    }
-
-    public static SavePointType get(String value) {
-        for (SavePointType type : SavePointType.values()) {
-            if (Asserts.isEqualsIgnoreCase(type.getValue(), value)) {
-                return type;
-            }
-        }
-        return SavePointType.TRIGGER;
-    }
-}
