--- dlink-gateway/src/main/resources/META-INF/services/com.dlink.gateway.Gateway
@@ -1,4 +1,0 @@
-com.dlink.gateway.yarn.YarnApplicationGateway
-com.dlink.gateway.yarn.YarnPerJobGateway
-com.dlink.gateway.kubernetes.KubernetesApplicationGateway
-com.dlink.gateway.kubernetes.KubetnetsApplicationOperatorGateway
