+++ pom.xml
@@ -90,16 +90,10 @@
         <knife4j.version>3.0.3</knife4j.version>
         <spotless.version>2.27.1</spotless.version>
         <jaxb.version>2.3.0</jaxb.version>
-        <flink.operator.version>1.2.0</flink.operator.version>
     </properties>
 
     <dependencyManagement>
         <dependencies>
-            <dependency>
-                <groupId>org.apache.flink</groupId>
-                <artifactId>flink-kubernetes-operator</artifactId>
-                <version>${flink.operator.version}</version>
-            </dependency>
             <!--解决mysql与phoenix依赖冲突-->
             <dependency>
                 <groupId>com.google.protobuf</groupId>
@@ -367,7 +361,17 @@
             </dependency>
             <dependency>
                 <groupId>com.dlink</groupId>
-                <artifactId>dlink-gateway</artifactId>
+                <artifactId>dlink-gateway-base</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.dlink</groupId>
+                <artifactId>dlink-gateway-yarn</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.dlink</groupId>
+                <artifactId>dlink-gateway-kubernets</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
@@ -506,6 +510,7 @@
                 <artifactId>activation</artifactId>
                 <version>1.1.1</version>
             </dependency>
+
         </dependencies>
     </dependencyManagement>
